+++ src/main/java/tn/esprit/spring/service/Interface/LeaveTestService.java
@@ -14,7 +14,7 @@ public interface LeaveTestService {
 	LeaveTest retrieveLeave(String id);
 	List<LeaveTest> retrieveLeaveByType(Type_Leave type);
 	List<LeaveTest> retrieveLeaveByReason(Leave_reason reason);
-	void affecterUserALeave(int idu, int id);
+	void affecterUserALeave(int idu, Long id);
 	public int duration();
 	List<Long> statmonthh();
 	List<String> statmonthhc();
